move emit_buf_to_tensor to mlir_gen utils
